amd_exporter: Prometheus exporter AMD E-SMI lib
This is the first commit of the GO language code for the
AMD EPYC Prometheus Exporter. The exporter is a systemd
daemon that facilitates the export of AMD specific CPU
data to the Prometheus Server. The exporter interfaces
with the AMD E-SMI library, using GO bindings defined
in the goesmi repo, to export CPU energy, power, and
boost limit information.

Change-Id: Ifc9bf96874fe88b557eb734a1b5ca597ebcfa8d1